requirements.txt
asgiref==3.6.0
azure-cognitiveservices-vision-customvision==3.1.0
azure-common==1.1.28
azure-core==1.26.3
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
psycopg2-binary==2.9.5
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.14
